Removed useCurrentExploreView hook, RootExplore and Explore containers handle their own views. Pairing with Amanda
