+++ src/components/Explore/RootExploreContainer.js
@@ -26,6 +26,8 @@ const RootExploreContainerWithContext = ( ): Node => {
   const navigation = useNavigation( );
   const { isConnected } = useNetInfo( );
   const currentUser = useCurrentUser( );
+  const rootExploreView = useStore( state => state.rootExploreView );
+  const setRootExploreView = useStore( state => state.setRootExploreView );
   const rootStoredParams = useStore( state => state.rootStoredParams );
   const setRootStoredParams = useStore( state => state.setRootStoredParams );
   const {
@@ -127,6 +129,8 @@ const RootExploreContainerWithContext = ( ): Node => {
         filterByIconicTaxonUnknown={
           () => dispatch( { type: EXPLORE_ACTION.FILTER_BY_ICONIC_TAXON_UNKNOWN } )
         }
+        currentExploreView={rootExploreView}
+        setCurrentExploreView={setRootExploreView}
         isConnected={isConnected}
         loadingStatus={loadingStatus}
         openFiltersModal={openFiltersModal}
